refactor: move the auth guard to the common folder
