+++ src/user/user.controller.ts
@@ -9,7 +9,7 @@ import {
 
 import { UserService } from './user.service';
 
-import { AuthGuard } from './guards/auth.guard';
+import { AuthGuard } from '../common/guards/auth.guard';
 
 import { Id } from './decorators/id.decorator';
